Updated Bishop, Queen and Rook
