+++ src/engine/pieces/rook.ts
@@ -10,18 +10,6 @@ export default class Rook extends Piece {
     }
 
     public getAvailableMoves(board: Board) {
-        let newPositions: Square[] = [];
-        const currentPosition= board.findPiece(this);
-        for (let i = 0; i < GameSettings.BOARD_SIZE; i++) {
-            if (i !== currentPosition.col) {
-                newPositions.push(new Square(currentPosition.row, i));
-            }
-        }
-        for (let i = 0; i < GameSettings.BOARD_SIZE; i++) {
-            if (i !== currentPosition.row) {
-                newPositions.push(new Square(i, currentPosition.col));
-            }
-        }
-        return newPositions;
+        return this.linearMove(board, board.findPiece(this));
     }
 }
